+++ setup.cfg
@@ -6,6 +6,12 @@ addopts =
     --cov=phoible
     --cov-report term-missing
 
+[coverage:run]
+source = phoible
+
+[coverage:report]
+show_missing = true
+
 [compile_catalog]
 directory = phoible/locale
 domain = clld
